sk/python/sk.ui/requirements.txt
streamlit
azure-storage-blob
requests
python-dotenv